Fix: monthlyPopularityScore 타입 int에서 double로 변경

See merge request s10-bigdata-recom-sub2/S10P22A306!226

+++ back/src/main/java/site/chachacha/fitme/domain/product/dto/ProductResponse.java
@@ -15,7 +15,7 @@ public class ProductResponse {
     private Integer price;
     private List<MainImageResponse> mainImages;
     private BrandResponse brand;
-    private int popularityScore;
+    private double popularityScore;
     private int likeCount;
     private double reviewRating;
     private int reviewCount;
